fix: adapta metodo processAttempt
