+++ termo/app/Attempt.hs
@@ -2,6 +2,29 @@ module Attempt where
 
 import Data.Char (toLower, toUpper)
 import Colors (red, reset)  -- Importa as variáveis red e reset de colors
+import System.Console.ANSI (setSGR, SGR(SetColor, Reset), ColorIntensity(Vivid), Color(..), ConsoleLayer(Foreground))
+import LetterStatus (LetterStatus(..))
+
+statusToColor :: LetterStatus -> Color
+statusToColor NoExist = Black
+statusToColor RightPlace = Green
+statusToColor WrongPlace = Yellow
+statusToColor _ = White 
+
+-- Função que imprime a letra com a cor correspondente
+printStyledLetter :: (Char, LetterStatus) -> IO ()
+printStyledLetter (ch, status) = do
+    setSGR [SetColor Foreground Vivid (statusToColor status)]
+    putStr [ch]
+    setSGR [Reset]
+
+-- Função que compara as letras e retorna o status
+compareLetters :: String -> Char -> Char -> (Char, LetterStatus)
+compareLetters secretWord secretLetter attemptLetter
+    | attemptLetter == secretLetter = (attemptLetter, RightPlace)
+    | attemptLetter `elem` secretWord = (attemptLetter, WrongPlace)
+    | otherwise = (attemptLetter, NoExist)
+
 
 -- Função para verificar a presença de uma letra e sua posição na palavra secreta
 containsElem :: Char -> String -> (Int, Int)
@@ -50,6 +73,7 @@ processAttempt :: String -> String -> IO Bool
 processAttempt secretWord attempt = do
     let lowercaseSecret = map toLower secretWord   -- Converte a palavra secreta para minúsculas
         lowercaseAttempt = map toLower attempt      -- Converte a tentativa para minúsculas
-        feedback = buildFeedback lowercaseSecret lowercaseAttempt
-    putStrLn feedback
-    return (lowercaseSecret == lowercaseAttempt)  -- Retorna True se a tentativa for igual à palavra secreta
+        statuses = zipWith (compareLetters lowercaseSecret) lowercaseSecret lowercaseAttempt
+    mapM_ printStyledLetter statuses                -- Imprime cada letra com o status colorido
+    putStrLn ""
+    return (lowercaseSecret == lowercaseAttempt)
